[CL-1744] Replace semantic ui tabe by new table in representativeness
