+++ front/app/modules/commercial/representativeness/admin/components/ChartCard/Table/HeaderRow.tsx
@@ -1,23 +1,30 @@
 import React from 'react';
 
 // components
-import { Table } from 'semantic-ui-react';
+import { Header, Row, HeaderCell } from 'components/admin/Table';
 import { Box } from '@citizenlab/cl2-component-library';
+
+// utils
+import { roundPercentage } from 'utils/math';
 
 interface Props {
   columns: string[];
 }
 
 const HeaderRow = ({ columns }: Props) => (
-  <Table.Header>
-    <Table.Row>
+  <Header>
+    <Row>
       {columns.map((column, i) => (
-        <Table.HeaderCell width={1} key={i}>
+        <HeaderCell
+          width={`${roundPercentage(1, columns.length)}%`}
+          key={i}
+          p="12px"
+        >
           <Box my="6px">{column}</Box>
-        </Table.HeaderCell>
+        </HeaderCell>
       ))}
-    </Table.Row>
-  </Table.Header>
+    </Row>
+  </Header>
 );
 
 export default HeaderRow;
