--- front/app/modules/commercial/representativeness/admin/components/ChartCard/Table/HeaderRow2.tsx
@@ -1,30 +1,0 @@
-import React from 'react';
-
-// components
-import { Header, Row, HeaderCell } from 'components/admin/Table';
-import { Box } from '@citizenlab/cl2-component-library';
-
-// utils
-import { roundPercentage } from 'utils/math';
-
-interface Props {
-  columns: string[];
-}
-
-const HeaderRow = ({ columns }: Props) => (
-  <Header>
-    <Row>
-      {columns.map((column, i) => (
-        <HeaderCell
-          width={`${roundPercentage(1, columns.length)}%`}
-          key={i}
-          p="12px"
-        >
-          <Box my="6px">{column}</Box>
-        </HeaderCell>
-      ))}
-    </Row>
-  </Header>
-);
-
-export default HeaderRow;
