+++ front/app/modules/commercial/representativeness/admin/components/ChartCard/Table/Table.tsx
@@ -1,33 +1,12 @@
 import React from 'react';
-import styled from 'styled-components';
 
 // components
-import { Table } from 'semantic-ui-react';
+import { Table, Body } from 'components/admin/Table';
 import HeaderRow from './HeaderRow';
 import Row from './Row';
 
-// styling
-import { colors } from 'utils/styleUtils';
-
 // typings
 import { RepresentativenessData } from '../../../hooks/createRefDataSubscription';
-
-const TABLE_HEADER_BG_COLOR = '#f9fafb';
-
-const StyledTable = styled(Table)<{ $hideBorderTop?: boolean }>`
-  td,
-  th > div {
-    color: ${colors.primary};
-  }
-  ${({ $hideBorderTop }) =>
-    $hideBorderTop ? 'border-top: 0px !important;' : ''}
-`;
-
-const StyledBody = styled(Table.Body)`
-  td:first-child {
-    background-color: ${TABLE_HEADER_BG_COLOR};
-  }
-`;
 
 interface Props {
   columns: string[];
@@ -36,14 +15,14 @@ interface Props {
 }
 
 const TableComponent = ({ columns, data, hideBorderTop }: Props) => (
-  <StyledTable $hideBorderTop={hideBorderTop}>
+  <Table borderTop={hideBorderTop ? 'none' : undefined}>
     <HeaderRow columns={columns} />
-    <StyledBody>
+    <Body>
       {data.map((row, i) => (
         <Row row={row} key={i} />
       ))}
-    </StyledBody>
-  </StyledTable>
+    </Body>
+  </Table>
 );
 
 export default TableComponent;
